+++ examples/SlideShow/slideShow.html
@@ -17,11 +17,15 @@
       const inputManager = new udv.Components.InputManager();
 
       app.start('../assets/config/all_widget_config.json').then((config) => {
-        app.addBaseMapLayer();
+        udv.Components.addBaseMapLayer(config, app.view, app.extent);
 
-        app.addElevationLayer();
+        udv.Components.addElevationLayer(config, app.view, app.extent);
 
-        app.setupAndAdd3DTilesLayers();
+        udv.Components.setupAndAdd3DTilesLayers(
+          config,
+          app.layerManager,
+          app.view
+        );
 
         const slideShow = new udv.Widgets.SlideShow(app, inputManager);
         app.addModuleView('slideShow', slideShow);
