+++ src/Components/Components.js
@@ -13,3 +13,11 @@ export { focusCameraOn } from './Camera/CameraUtils';
 
 export { InputManager } from './InputManager';
 
+export { LayerManager } from './LayerManager/LayerManager.js';
+
+import * as THREEUtils from './THREEUtils';
+export { THREEUtils };
+
+export { setupAndAdd3DTilesLayers } from './ViewUtils';
+export { addBaseMapLayer } from './ViewUtils';
+export { addElevationLayer } from './ViewUtils';
