+++ src/Components/InputManager.js
@@ -1,7 +1,5 @@
 /** @format */
 
-const Game = require('../Game/Game');
-
 /**
  * Poll system (https://en.wikipedia.org/wiki/Polling_(computer_science))
  * Inputs state is stored asynchronously then state is access synchronously by user
