+++ src/Components/LayerManager/Docs/layerManager.md
