+++ src/Game/Components/Components.js
@@ -1,14 +1,12 @@
 /** @format */
 
 const commonJsPack = require('./Pack');
-const commonJsTHREEUtils = require('./THREEUtils');
 const commonJsRoutine = require('./Routine');
 const commonJsJSONUtils = require('./JSONUtils');
 const commonJsType = require('./Type');
 
 module.exports = {
   Pack: commonJsPack,
-  THREEUtils: commonJsTHREEUtils,
   Routine: commonJsRoutine,
   JSONUtils: commonJsJSONUtils,
   Type: commonJsType,
