+++ src/Templates/Templates.js
@@ -1,4 +1,3 @@
 /** @format */
-import { LocalGame } from './Game/LocalGame/LocalGame';
-export { LocalGame };
+export { LocalGame } from './Game/LocalGame/LocalGame';
 export { AllWidget } from './AllWidget/AllWidget';
