+++ src/Views/Billboard/Billboard.js
@@ -1,7 +1,7 @@
 /** @format */
 
 import * as THREE from 'three';
-import THREEUtils from '../../Game/Components/THREEUtils';
+import * as THREEUtils from '../../Components/THREEUtils';
 import { CSS3DObject } from 'three/examples/jsm/renderers/CSS3DRenderer';
 import { Color } from 'three';
