+++ src/Widgets/CityObjects/ViewModel/CityObjectProvider.js
@@ -7,7 +7,7 @@ import {
   CityObject,
 } from '../../../Components/3DTiles/Model/CityObject';
 import { EventSender } from '../../../Components/Events/EventSender';
-import { LayerManager } from '../../Components/LayerManager/LayerManager';
+import { LayerManager } from '../../../Components/LayerManager/LayerManager';
 
 import { CityObjectFilter } from './CityObjectFilter';
 import { CityObjectLayer } from './CityObjectLayer';
