+++ src/Widgets/Components/Components.js
@@ -2,7 +2,5 @@
 
 export { ModuleView } from './ModuleView/ModuleView.js';
 
-export { LayerManager } from './LayerManager/LayerManager.js';
-
 import * as GUI from './GUI/GUI.js';
 export { GUI };
